Data.Dependency

Description

Dependant streams of values.

A structure that can be thought of as producing many values, but optionally (but explicitly!) depending on some other.

This lets use the same syntax for creating structures that do or do not depend on some outside data, while still being able to inspect those structures that don't ultimately need the dependency. It is trivially isomorphic to stream types used by the streaming library and heavily inspired by that, with the main difference being that the functions here allow for streams to be inspected without necessarily going through the underlying stream monad - see pop.

Base type

newtype Stream m a

A version of StreamM with Functor and Foldable instances over the elements of the stream rather than the inner results of the monad

Constructors

Stream

Fields

unStream :: StreamM m a ()

data StreamM m a x

A stream type that can be inspected without always needing to first go through the m effect. This is in contrast to the streams provided by the "streaming" package that _always_ require going through the effect monad to unwrap a stream. It also supports cheap concatentation of streams.

The property of being able to inspect the stream without going through its effect monad is that this stream type is very useful for modelling "dependant" streams, where some amount of data can be generated in staggered stages, contingent on explicitly-made requests for more input.

Constructors

Got a (StreamM m a x)
A yielded value of the stream

Wait (m (StreamM m a x))
Wait for a value

Cat (StreamM m a x) (StreamM m a x)
Concatenate streams

Done x
Monadic return

Constructing streams

yield :: a -> StreamM m a ()

lift :: Functor m => m x -> StreamM m a x

Inspecting streams

data Pop m a x

Constructors

Pop a (StreamM m a x)

PopWait (m (Pop m a x))

PopDone x

pop :: Functor m => StreamM m a x -> Pop m a x

Pop the first available element of the stream, rebalancing the stream if possible so that subsequent pop calls are cheap. When PopDone is reached, the stream has ended. The reason that this is given its own type instead of immediately returning m (a, StreamM m a x) is because many streams in this project are in fact pure. So the actual type should be like Either (Maybe (a, Stream m a x)) (m (Maybe (a, StreamM m a x))) which is obviously much more unwieldy; hence a custom type.

popForced :: Monad m => StreamM m a x -> m (Maybe (a, StreamM m a x))

Convenience

forcePop :: Monad m => Pop m a x -> m (Maybe (a, StreamM m a x))

Force a Pop, using the monad even if it is not strictly necessary.

peek :: StreamM m a x -> Maybe (a, StreamM m a x)

Purely pop off the first available element of the stream, rebalancing the stream so that subsequent uncons operations are cheap. If the ordering of elements and effects is important for this stream, prefer uncons instead -- this function is happy to pop off a yielded stream element from the middle of stream.

peekFoldMap :: Monoid b => (a -> b) -> StreamM m a x -> b

Fold over only the immediately available elements

peekAll :: StreamM f a x -> (Vector a, Bool, StreamM f a x)

Purely pop off all available elements of the stream. The ordering of elements may not be the order that is expected if there are effects that appear in the tree "before" available elements.

peekAllS :: MonadState (StreamM f a x) m => m (Bool, Vector a)

null :: StreamM f a x -> Bool

Check whether a stream has any remaining elements or effects

Mapping streams

map :: Functor m => (a -> b) -> StreamM m a x -> StreamM m b x

hoist :: Functor f => (forall x. f x -> g x) -> StreamM f a r -> StreamM g a r

Natural transformation over the inner monad of a stream.

hoistT :: (MonadTrans t, Monad m) => StreamM m a r -> StreamM (t m) a r

Natural transformation over the inner monad of a stream.

for :: Monad m => StreamM m a x -> (a -> m b) -> StreamM m b x

Iterate over the stream, running effects as necessary.

for_ :: Monad m => StreamM m a x -> (a -> m ()) -> m ()

Iterate over the stream, running effects as necessary.

fold :: Monad m => (a -> r -> r) -> r -> StreamM m a x -> m r

Strictly fold over the elements of the stream, running effects as needed

foldM :: Monad m => (a -> r -> m r) -> r -> StreamM m a x -> m r

Strictly fold over the elements of the stream, running effects as needed

ifoldM :: Monad m => (Int -> a -> r -> m r) -> r -> StreamM m a x -> m r

Strictly fold over the elements of the stream, running effects as needed

Composing streams

alignWith :: forall m a b c. Monad m => (These a b -> c) -> StreamM m a () -> StreamM m b () -> StreamM m c ()

Alignment of two streams. Avoids running effects until necessary.

data Group a

Constructors

GroupSplit !a !a

GroupCombine !a

group :: forall a m x. Monad m => (a -> a -> m (Group a)) -> StreamM m a x -> StreamM m a ()

Group consecutive elements of a stream

groupr :: forall m c b a x. (Monoid b, Functor m) => (a -> b) -> (b -> c) -> StreamM m a x -> StreamM m c x

Optimise the representation of a stream by grouping consecutive pure elements through a monoid, and finalising that at the end

Reducing streams to vectors or lists

toVector :: Monad m => StreamM m a x -> m (Vector a)

Run a stream and its effects, collecting the results into a vector

toList :: Monad m => StreamM m a x -> m [a]

Run a stream and its effects, collecting the results into a list

pureVector :: StreamM m a x -> Vector a

Collect the pure elements only of a stream into a vector

pureList :: StreamM m a x -> [a]

Collect the pure elements only of a stream

Request helpers

once :: (MonadIO outer, MonadIO inner, NFData a) => inner a -> outer (inner a)

Perform an action only once. This may be useful for ensuring a dependency only is fetched once.

cached :: (MonadState s m, NFData a) => Lens' s (Maybe a) -> m a -> m a

Pure MonadState version of once, that takes a lens to the cached value from the state, and a function to compute a value.