Tourney.Algebra.Builder

Documentation

data Builder t r a

Overloaded constructors

Steps

class AsSteps a r where

Values that have simple Steps representations. Think of this as an overloaded Steps constructor.

Methods

steps :: a -> Steps r ()

type Steps = Builder TMany

A monad to build tournaments over steps (See (***)). Operations in this monad create rounds. See Round for a builder monad to create rounds.

round_ :: forall a m. (MonadRounds m, AsRound a ()) => a -> m ()

Add a round to steps

rounds_ :: AsRound r () => [r] -> Steps s ()

Add some rounds to steps

asRound :: Round r a -> Round r a

Rounds

class AsRound a r where

Methods

toRound :: a -> Round r ()

type Round = Builder TOne

A monad to build tournaments over matches. For convenience, this monad behaves like a list monad, where the inner list is ultimately combined by overlay.

match :: AsMatch m => m -> Round r ()

Add a match to the round

Matches

class AsMatch m where

Methods

toMatch :: m -> Match

Basic syntax

list :: Foldable f => f a -> Builder t r a

Lift a list into a Round for iteration. The underlying effect on the Round is to overlay the iterated elements

overlaysBy :: (Foldable f, Merge t) => (a -> x) -> f (Builder t x a) -> Builder t r [x]

Overlay a collection of builders

overlays :: (Foldable f, Merge t) => f (Builder t a a) -> Builder t r [a]

Overlay a collection of builders

overlays_ :: (Foldable f, Merge t) => f (Builder t a a) -> Builder t r ()

Overlay a collection of builders

getStandings :: Merge t => Builder t () Standings

Retrieve the current standings

Warning: this operation will have the effect of segmenting the tournament into two sections, as a tournament runner cannot continue past this point without first having the standings.

getPlayerCount :: Merge t => Builder t () PlayerCount

Retrieve the current player count

withFocus :: Merge t => Slot -> Int -> Builder t a a -> Builder ('TMod t) r ()

Set the focus of the inner builder

withFocii :: Merge t => (Focus -> [Focus]) -> Builder t a a -> Builder ('TMod t) r ()

Set the focii of the inner builder

withOffset :: Merge t => Int -> Builder t a a -> Builder ('TMod t) r ()

Set the offset of the inner builder

divideInto :: Int -> Steps () () -> Steps () ()

Given a denominator n and a builder, split the tournament into n sub-tournaments and run the builder on each

foldAround :: Int -> [Slot] -> [Match]

Fold a list of players together around a midpoint of the list.

foldAroundMidpoint :: [Slot] -> [Match]

Fold a list of players together around _the_ midpoint of the list. You should ensure the input list has an even number of players.

swaps :: Merge t => Builder t a a -> Builder t r ()

points :: Merge t => Builder t () () -> Builder t r ()

barrier :: Merge t => Builder t () () -> Builder t r ()

Reduction back into a Tournament

inspect :: (Merge t, Merge i) => Inspect i a -> Builder i r r -> Builder t () a

Inspect the inner builder, for, e.g., matches

execBuilder :: Merge t => (a -> r) -> Builder t r a -> Tournament t

execSteps :: (a -> r) -> Steps r a -> Tournament TMany

execRound :: (a -> r) -> Round r a -> Tournament TOne

runSteps :: (a -> r) -> Steps r a -> ([r], Tournament TMany)

runRound :: (a -> r) -> Round r a -> ([r], Tournament TOne)

getAccum :: Merge t => Builder t r (Tournament t)