Tourney.Algebra.Unified

Base eDSL

The core primitives and syntax of the journeyman tournament eDSL are defined here. For how these values can be transformed into a flattened stream of rounds, see Tourney.Stream

data Tournament :: Depth -> Type where

A description of a tournament. The type-parameter gives an upper-bound on how many rounds this tournament has. It also ensures that the spine of a tournament is static, regardless of use of constructors such as ByStandings, that enable creating tournaments that depend on match results that are not yet known. Intuitively if you have a tournament with depth TMany, it means it contains multiple rounds, whereas if you have a tournament with depth TOne, then it is only one round.

Constructors

One :: Match -> Tournament 'TOne

Empty :: Tournament t

Modify :: Mod -> Tournament t -> Tournament (TMod t)
Modify a tournament by some Mod, such as setting the focus.

Overlay :: Tournament a -> Tournament a -> Tournament a
Overlay two tournaments, to describe running two sub-tournaments in parallel. The depth of the tournaments must be the same

Sequence :: (KnownDepth a, KnownDepth b) => Tournament a -> Tournament b -> Tournament TMany
Sequence two tournaments one after the other. The resulting tournament has a depth TMany which restricts what functions are able to manipulate it.

Sort :: SortMethod -> Tournament t -> Tournament t
Sort the inner tournament by some sorting method

ByPlayerCount :: (PlayerCount -> Tournament t) -> Tournament t
Depend on the player count to produce an inner tournament

ByStandings :: (Standings -> Tournament t) -> Tournament t
Depend on the current standings, at the outset of the current round, to run the tournament.

LiftTOne :: Tournament TOne -> Tournament TMany
Lift a single round of a tournament into having a depth TMany

LiftTMod :: KnownDepth t => Tournament (TMod t) -> Tournament TMany
Lift a modified round of a tournament into having a depth TMany

data Mod

Constructors

SetFocus !(Focus -> [Focus])

SetOffset !Int

Operators

Overlays of tournaments

The verb overlay comes from algebraic graph theory, and means to combine two graphs by simply having them coexist in the same graph; no actual edges are drawn.

(+++) :: Tournament a -> Tournament a -> Tournament a infixl 2

Overlay two tournaments. See Overlay.

overlay :: Foldable f => f (Tournament a) -> Tournament a

Overlay an arbitrary number of tournaments.

Sequencing tournaments

Sequencing means to connect two tournaments by having one happen before the other. A round means a single step of a tournament such as a group of matches.

(***) :: forall a b. (KnownDepth a, KnownDepth b) => Tournament a -> Tournament b -> Tournament TMany infixl 1

Sequence two tournaments, running them one after the other. See Sequence.

sequence :: (Foldable f, KnownDepth a) => f (Tournament a) -> Tournament TMany

Extra Types

data SortMethod

Constructors

WinnerTakesHigh

PointsAward
Award points and use that to sort the results in the end

data Sorter

Constructors

Sorter

Fields

sorterFocus :: !Focus

sorterMethod :: !SortMethod

data Focus

The focus of a tournament over a sorting network. A good intuition is that a Focus is basically an array slice over a round.

Constructors

Focus

Fields

focusStart :: !Slot

focusLength :: !Int

data Depth

The depth of a tournament. Since we only care about distinguishing single rounds from sequences of rounds, that is the only information stored here.

Constructors

TOne

TMany

TMod Depth

class Typeable d => KnownDepth (d :: Depth) where

Methods

depthVal :: proxy d -> Depth