Tourney.Common

Description

Common types and functions

Types

Type-safe wrappers around numeric types

This prevents mixing up two different ideas, namely Slot and Player, due to their representations being the same.

type PlayerCount = Int

newtype Player

The basic type for a player is an Int denoting the "nth" best player in the tournament at the outset of it. A Player is the data that gets sorted by the sorting networks constructed by tournaments.

Constructors

Player Int

newtype Slot

A slot denotes an index into a round of a tournament, which a Player occupies.

Constructors

Slot Int

fromSlot :: Slot -> Int

newtype RoundNo

Round number

Constructors

RoundNo Int

Standings

data Standings

A mapping from slots to players

createInitialStandings :: PlayerCount -> Standings

modifyStandings :: Standings -> (forall s. MVector s Player -> ST s ()) -> Standings

vectorToStandings :: Vector v Player => v Player -> Standings

Conversions

asInt :: Coercible a Int => a -> Int

Ranges

(..<) :: (Enum a, Num a) => a -> a -> [a] infix 5

Create an increasing [closed, open) interval of slots

(..>) :: (Enum a, Num a) => a -> a -> [a] infix 5

Create a decreasing [closed, open) interval of slots

Sort methods

data SortMethod

Constructors

WinnerTakesHigh

PointsAward
Award points and use that to sort the results in the end

data Sorter

Constructors

Sorter

Fields

sorterFocus :: !Focus

sorterMethod :: !SortMethod

Focus

data Focus

The focus of a tournament over a sorting network. A good intuition is that a Focus is basically an array slice over a round.

Constructors

Focus

Fields

focusStart :: !Slot

focusLength :: !Int

focusEnd :: Focus -> Slot

focusWithin :: Focus -> Focus -> Bool

focusContains :: Focus -> Slot -> Bool

Arithmetic functions

Where appropriate these are overloaded to work with types that have the same representation as Int, such as Player, Slot, and RoundNo.

stride2 :: Int -> OrdPair Int

Expand an index to 2 dimensions

nearestPow2Above :: Coercible Int a => a -> a

Round to the nearest power of 2 above the input

bitLog2 :: Coercible Int a => a -> a

Integer log2

isqrt :: Coercible Int a => a -> a

Integer square root

triangular :: Coercible Int a => a -> a

Generate the N-th triangular number