Tourney.Match.Matrix

Description

Sparse matrices for holding matches

Documentation

data MatchMatrix

An abstraction around a mutable store of rounds of matches and their results. Conceptually it stores a matrix of matches by (round, player1, player2), and a variable for the current round.

newtype MapByRound a

Constructors

ByRound (IntMap a)

vectorMapByRound :: Vector a -> MapByRound a

newtype MapByMatches a

A sparse matrix for storing a map between matches and a value.

Since the two players in Match are guaranteed to be ordered (low, high) we only need to ever check the low->high->match path in the nested intmaps.

Constructors

ByMatches (IntMap (IntMap a))

createMatchMatrix :: PlayerCount -> STM MatchMatrix

readMatchMatrix :: MatchMatrix -> STM (MapByRound (MapByMatches (Maybe Result)))

getMatch :: MatchMatrix -> RoundNo -> Match -> STM (Maybe (Maybe Result))

Get a match

getMatches :: MatchMatrix -> RoundNo -> STM (Vector (Match, Maybe Result))

getMatchResults :: MatchMatrix -> RoundNo -> STM (Vector MatchResult)

addMatch :: MatchMatrix -> RoundNo -> Match -> STM ()

Add a match. It must not exist already in the current round

setMatchResult :: MatchMatrix -> RoundNo -> Match -> Result -> STM ()

Set a match result. It must already exist in the current round, by addMatch

getPendingMatches :: MatchMatrix -> RoundNo -> STM (Vector Match)

pendingMatchCount :: MatchMatrix -> RoundNo -> STM Int

haveAnyPendingMatchesWithin :: MatchMatrix -> RoundNo -> Focus -> STM Bool