Tourney.UI.Main

Documentation

knownTournaments :: [(Text, Tournament TMany)]

knownTournamentsLen :: Int

data MenuForm

Constructors

MenuForm

Fields

tournament :: !Int

playerCount :: !Text

data AppState

Constructors

AppState

Fields

menu :: Form MenuForm AppEvent AppResourceName

dialog :: Dialog DialogChoice AppResourceName

state :: !(Maybe TournamentState)

errors :: [Text]

data TournamentState

Constructors

TournamentState

Fields

tournament :: Tournament TMany

selection :: !Selection

matches :: MapByRound (MapByMatches (Maybe Result))

roundNo :: !RoundNo

pendingMatches :: Vector Match

pureMatches :: MapByRound (MapByMatches (Maybe Result))

pureCode :: Code

codeSoFar :: Vector (TourneyOp, StepCodeEvent)

rawEvents :: Vector StepCodeEvent

standingsHistory :: MapByRound StandingsUpdate

vm :: !VM

playerCount :: !PlayerCount

colours :: !(Array Player (Colour Double))

tournamentViewportScroll :: !(Int, Int)

matchForm :: Maybe (Form MatchForm AppEvent AppResourceName)

data MatchForm

Constructors

MatchForm

Fields

score1 :: !Text

score2 :: !Text

data AppResourceName

Constructors

Tournament

CodeStream

EventStream

Panel

ScrollBar ClickableScrollbarElement AppResourceName

MenuTournamentItem !Int

MenuPlayerCountItem

MenuStandingsItem

MenuEnter

Score1

Score2

type UIElement = Reader TournamentState (Widget AppResourceName)

data AppEvent

main :: IO AppState

data DialogChoice

Constructors

DialogEnter

menuDialog :: Dialog DialogChoice AppResourceName

menuForm :: MenuForm -> Form MenuForm AppEvent AppResourceName

drawMenu :: Reader AppState (Widget AppResourceName)

beginTournament :: EventM AppResourceName AppState ()

drawErrors :: [Text] -> [Widget AppResourceName]

handleTournamentEvent :: BrickEvent AppResourceName AppEvent -> EventM AppResourceName TournamentState ()

createTournamentState :: PlayerCount -> Tournament TMany -> IO TournamentState

advanceRound :: EventM AppResourceName TournamentState ()

quickAssignResult :: Result -> EventM AppResourceName TournamentState ()

acceptMatchFormEntry :: EventM AppResourceName TournamentState ()

parsePoints :: Text -> MaybeT (EventM AppResourceName s) Points

createMatchForm :: EventM AppResourceName TournamentState (Maybe MatchForm)

drawMain :: Reader TournamentState [Widget AppResourceName]

niceBorder :: Widget a -> Widget a

drawCode :: UIElement

drawEvents :: UIElement

drawPendings :: UIElement

drawSelectedMatch :: Reader TournamentState (Maybe (Widget AppResourceName))

matchForm :: MatchForm -> Form MatchForm AppEvent AppResourceName

drawHelp :: UIElement

helpText :: Text -> Widget n

drawTournament :: UIElement

codeRealAttr :: AttrName

codePureAttr :: AttrName

helpAttr :: AttrName

helpKeyAttr :: AttrName

stepCodeAttr :: AttrName

lostPlayerAttr :: AttrName

winPlayerAttr :: AttrName

rowOddAttr :: AttrName

rowEvenAttr :: AttrName

dim :: Color

dimRed :: Color

makeMatchColours :: PlayerCount -> Array Player (Colour Double)

toVtyRGB :: Colour Double -> Color

getMatchAttr :: PlayerCount -> Array Player (Colour Double) -> (Slot, Maybe Player) -> (Slot, Maybe Player) -> Slot -> Attr

chunksOf :: Int -> [e] -> [[e]]

chunksOf function taken from the "split" package by Brent Yorgey 2008-2023 licensed BSD-3

Copyright (c) 2008 Brent Yorgey, Louis Wasserman

All rights reserved.

Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met: 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer. 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution. 3. Neither the name of the author nor the names of other contributors may be used to endorse or promote products derived from this software without specific prior written permission.

THIS SOFTWARE IS PROVIDED BY THE AUTHORS ``AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE AUTHORS OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.