Tourney.VM

Documentation

data VM

An encapsulation of both a compiler and an interpreter of tournaments.

Core VM functionality

setup :: Tournament t -> PlayerCount -> IO VM

Create a VM from a tournament and a fixed player count.

loop :: VM -> IO (Vector StepCodeEvent)

Evaluate a VM to the maximal extent allowed. Inspecting the result of this function will reveal why the tournament stopped when it did; for instance, if there are still pending matches, the final value in the StepCodeEvent vector will be Stepped ('NeedResults'Sorting' _ matches) or 'NeedResults'EndRound' _ matches

Functions for interacting with a VM

data StepCodeEvent

An event generated by stepping the VM

Constructors

NoCode

Stepped StepContinue

data StepContinue

Constructors

NeedResults'Sorting !RoundNo !(Vector Match)

NeedResults'EndRound !RoundNo !(Vector Match)

Continue !StepInfo

getCodeSoFar :: VM -> IO Code

getMatches :: VM -> IO (MapByRound (MapByMatches (Maybe Result)))

getStandingsHistory :: VM -> IO (MapByRound StandingsUpdate)

getPendingMatches :: VM -> IO (Vector Match)

getRoundNo :: VM -> IO RoundNo

peekCode :: VM -> IO Code

setMatchResult :: VM -> RoundNo -> Match -> Result -> IO Bool

getMatch :: VM -> RoundNo -> Match -> IO (Maybe (Maybe Result))

Re-exports

type Code = Vector TourneyOp

data TourneyOp

Constructors

MATCH !Match

BEGIN_ROUND
Begin a round

END_ROUND
End a round

PERFORM_SORTING !Focus !SortMethod
Cause the current sorting network to evaluate

newtype MapByRound a

Constructors

ByRound (IntMap a)

newtype MapByMatches a

A sparse matrix for storing a map between matches and a value.

Since the two players in Match are guaranteed to be ordered (low, high) we only need to ever check the low->high->match path in the nested intmaps.

Constructors

ByMatches (IntMap (IntMap a))

data Match

The basic type for a match is a pair of players. The Match type ensures by construction that the "lower" player always takes the first slot of the Match, and the higher the second. Many functions in this library are overloaded so that you can represent matches as ordinary Haskell tuples.

It is a runtime error to construct a match using the same player twice.

The first field of a Match is always the smaller value. Matches with one invalid Slot for any reason should be interpretted as a bye for the valid Slot in that match.

Constructors

Match_ !Slot !Slot

Bundled Patterns

pattern Match :: Slot -> Slot -> Match

data Result

Constructors

Result !Points !Points

data MatchResult

Constructors

MatchResult

Fields

match :: !Match

result :: !Result

data StandingsUpdate

Constructors

StandingsUpdate

Fields

roundNo :: !RoundNo

roundDepth :: !Int

standings :: !(Vector (Points, Player))